refactor: finish extension layered layout
