+++ extension/src/app/application/overlay/file-view.ts
@@ -1,5 +1,10 @@
-import { type BackgroundError, type DiffV2, type SemanticDiffResponse, unresolvedRemaining } from "../../domain/diff/types";
-import { must } from "../../presentation/util/must";
+import {
+  type BackgroundError,
+  type DiffV2,
+  type SemanticDiffResponse,
+  unresolvedRemaining,
+} from "../../domain/diff/types";
+import { must } from "../../domain/must";
 import type { View } from "./view-mode";
 
 // Per-file raw/semantic state machine (host + fetch latch); unit-testable without a browser.
