+++ extension/src/app/domain/unity.parity.test.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { expect, it } from "vitest";
-import { must } from "../presentation/util/must";
+import { must } from "./must";
 
 // The UnityYAML extension prefilter is hand-copied in two places: the CLI's
 // git-side gate (unity_path.zig) and the extension's path check (unity.ts).
