+++ extension/src/app/infrastructure/providers/wasm-differ.ts
@@ -2,7 +2,6 @@ import { DiffError, type DifferPort } from "../../application/port/differ";
 import type { DiffErrorV1, DiffV2 } from "../../domain/diff/types";
 
 export { DiffError };
-export type Differ = DifferPort;
 
 type Exports = {
   memory: WebAssembly.Memory;
